Crossfilter post-v4 code cleanup
Now that the d3v4 version works as well as the d3v3 version, remove all
the old d3v3 code and remove all _new references.  Yay
