+++ crossfilter_example/index.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset='utf-8'>
+        <title></title>
+
+        <link rel="stylesheet" href="worm_filter.css">
+        <script src="crossfilter_parameters.js"></script>
+        <script src="crossfilter.min.js" defer></script>
+        <script src="d3.v4.min.js" defer></script>
+        <script src="babel.min.js" defer></script>
+        <script src="crossfilter_bar_chart.js" defer></script>
+        <script src="crossfilter_helpers.js" defer></script>
+        <script src="crossfilter_demonstration.js" defer></script>
+    </head>
+</html>
+<body>
+    <div id="header">
+        <h1 id="crossfilter_report_title"></h1>
+    </div>
+    <div>
+        <a href="javascript:resetAll()" class="reset" style="margin-left: -10px; margin-bottom: 20px">Reset all</a>
+    </div>
+    <div id='charts'>
+        <div id='chart0' class='chart'>
+            <div class='title'></div>
+        </div>
+        <div id='chart1' class='chart'>
+            <div class='title'></div>
+        </div>
+        <div id='chart2' class='chart'>
+            <div class='title'></div>
+        </div>
+        <div id='chart3' class='chart'>
+            <div class='title'></div>
+        </div>
+    </div>
+
+    <div id="daySelectionDiv"></div>
+    <div id="datasetview">
+        <div class="title"></div>
+        <canvas></canvas>
+    </div>
+
+    <aside id='totals'>
+        <span id='active'>-</span> of <span id='total'>-</span> experiments selected.
+    </aside>
+
+    <div id='lists'>
+        <div id='results-list' class='result_row_list'></div>
+    </div>
+</body>
