--- crossfilter_example/old_v3_version/index.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>Worm Crossfilter Demonstration</title>
-    
-        <link rel="stylesheet" href="worm_filter.css">
-        <script src="crossfilter_parameters.js"></script>
-    </head>
-    
-    <body>
-        <div id="header">
-            <h1 id="report_title"></h1>
-        </div>
-        <div id="body">
-          <div>
-            <a href="javascript:resetAll()" class="reset" style="margin-left: -10px; margin-bottom: 20px">Reset all</a>
-          </div>
-          <div id="charts">
-            <div id="hour-chart" class="chart">
-              <div class="title">Time of Day</div>
-            </div>
-            <div id="delay-chart" class="chart">
-              <div class="title">Worm length (microns)</div>
-            </div>
-            <div id="distance-chart" class="chart">
-              <div class="title">Distance travelled (microns)</div>
-            </div>
-            <div id="date-chart" class="chart">
-              <div class="title">Date</div>
-            </div>
-          </div>
-        
-          <div id="daySelectionDiv"></div>
-        
-          <div id="datasetview">
-              <div class="title"></div>
-              <canvas></canvas>
-          </div>
-        
-          <aside id="totals"><span id="active">-</span> of <span id="total">-</span> experiments selected.</aside>
-
-          <div id="wormVisualization">
-              <svg></svg>
-          </div>
-          
-          <div id="lists">
-            <div id="results-list" class="list"></div>
-          </div>
-        </div>
-    
-        <script src="crossfilter.min.js"></script>
-        <script src="https://d3js.org/d3.v3.min.js"></script>
-        <script src="crossfilter_bar_chart.js"></script>
-        <script src="crossfilter_helpers.js"></script>
-        <script src="worm_visualization.js"></script>
-        <script src="crossfilter_demonstration.js"></script>
-    </body>
-</html>
